refactor alarm service package
