+++ app/src/main/AndroidManifest.xml
@@ -17,8 +17,8 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <receiver android:name=".schedule.alarm.service.AlarmReceiver" />
-        <service android:name=".schedule.media.AudioPlayerService" />
+        <receiver android:name=".alarm.service.AlarmReceiver" />
+        <service android:name=".alarm.media.AudioPlayerService" />
 
         <activity
             android:name=".schedule.ui.activity.ScheduleListActivity"
@@ -67,7 +67,7 @@
                 android:value="com.github.xzwj87.todolist.schedule.ui.activity.ScheduleListActivity" />
         </activity>
 
-        <activity android:name=".schedule.ui.activity.AlarmAlertActivity"
+        <activity android:name=".alarm.ui.activity.AlarmAlertActivity"
             android:theme="@style/AlarmAlertDialog">
         </activity>
