+++ app/src/main/java/com/github/xzwj87/todolist/alarm/shake/IShakeListener.java
@@ -1,4 +1,4 @@
-package com.github.xzwj87.todolist.schedule.shake;
+package com.github.xzwj87.todolist.alarm.shake;
 
 /**
  * Created by JasonWang on 2016/3/28.
